+++ nocropper/src/main/java/com/fenchtose/nocropper/CropState.java
@@ -1,0 +1,8 @@
+package com.fenchtose.nocropper;
+
+public enum CropState {
+    STARTED,
+    SUCCESS,
+    ERROR,
+    FAILURE_GESTURE_IN_PROCESS
+}