+++ nocropper/src/main/java/com/fenchtose/nocropper/CropperImageView.java
@@ -46,8 +46,6 @@ public class CropperImageView extends ImageView {
     private boolean mFirstRender = true;
 
     private Bitmap mBitmap;
-    private boolean doPreScaling = false;
-    private float mPreScale;
 
     private boolean mAddPaddingToMakeSquare = true;
 
@@ -248,23 +246,12 @@ public class CropperImageView extends ImageView {
             return;
         }
 
-        if (bm.getHeight() > 1280 || bm.getWidth() > 1280) {
+        if (bm.getHeight() > 1280 || bm.getWidth() > 1280 && DEBUG) {
             Log.w(TAG, "Bitmap size greater than 1280. This might cause memory issues");
         }
 
         mBitmap = bm;
-
-        if (doPreScaling) {
-            int max_param = Math.max(bm.getWidth(), bm.getHeight());
-            mPreScale = (float) max_param / (float) getWidth();
-
-            Bitmap scaledBitmap = Bitmap.createScaledBitmap(bm, (int) (bm.getWidth() / mPreScale),
-                    (int) (bm.getHeight() / mPreScale), false);
-            super.setImageBitmap(scaledBitmap);
-        } else {
-            mPreScale = 1f;
-            super.setImageBitmap(bm);
-        }
+        super.setImageBitmap(bm);
         requestLayout();
     }
 
@@ -594,14 +581,6 @@ public class CropperImageView extends ImageView {
         return getMatrixValue(matrix, Matrix.MSCALE_X);
     }
 
-    public boolean isDoPreScaling() {
-        return doPreScaling;
-    }
-
-    public void setDoPreScaling(boolean doPreScaling) {
-        this.doPreScaling = doPreScaling;
-    }
-
     public float getMaxZoom() {
         return mMaxZoom;
     }
@@ -693,7 +672,7 @@ public class CropperImageView extends ImageView {
 
     }
 
-    private Bitmap getCroppedBitmap() throws OutOfMemoryError {
+    public CropInfo getCropInfo() {
         if (mBitmap == null) {
             Log.e(TAG, "original image is not available");
             return null;
@@ -701,189 +680,113 @@ public class CropperImageView extends ImageView {
 
         Matrix matrix = getImageMatrix();
 
-        if (doPreScaling) {
-            matrix.postScale(1/mPreScale, 1/mPreScale);
-        }
-
         float xTrans = getMatrixValue(matrix, Matrix.MTRANS_X);
         float yTrans = getMatrixValue(matrix, Matrix.MTRANS_Y);
         float scale = getMatrixValue(matrix, Matrix.MSCALE_X);
 
         if (DEBUG) {
             Log.i(TAG, "xTrans: " + xTrans + ", yTrans: " + yTrans + " , scale: " + scale);
-        }
-
-        // Load bitmap which is mutable
-        Bitmap bitmap = null;
-
-        if (DEBUG) {
             Log.i(TAG, "old bitmap: " + mBitmap.getWidth() + " " + mBitmap.getHeight());
         }
 
+        // No scale/crop required.
         if (xTrans > 0 && yTrans > 0 && scale <= mMinZoom) {
-            // No scale/crop required.
             // Add padding if not square
-            if (mAddPaddingToMakeSquare) {
+            int verticalPadding = mBitmap.getHeight() > mBitmap.getWidth() ? 0 : (mBitmap.getWidth() - mBitmap.getHeight()) / 2;
+            int horizontalPadding = mBitmap.getWidth() > mBitmap.getHeight() ? 0 : (mBitmap.getHeight() - mBitmap.getWidth()) / 2;
+            return CropInfo.cropCompleteBitmap(mBitmap, mAddPaddingToMakeSquare, horizontalPadding, verticalPadding, mPaintColor);
+        }
 
-                try {
-                    return BitmapUtils.addPadding(mBitmap, mPaintColor);
-                } catch (OutOfMemoryError e) {
-                    Log.d(TAG, "OOM while adding padding");
-                    throw e;
-                }
+        float cropY = - yTrans / scale;
+        float Y = getHeight() / scale;
+        float cropX = -xTrans / scale;
+        float X = getWidth() / scale;
 
-            } else {
-                return mBitmap;
-            }
+        if (DEBUG) {
+            Log.i(TAG, "cropY: " + cropY);
+            Log.i(TAG, "Y: " + Y);
+            Log.i(TAG, "cropX: " + cropX);
+            Log.i(TAG, "X: " + X);
+        }
 
-        } else {
-
-            float cropY = - yTrans / scale;
-            float Y = getHeight() / scale;
-            float cropX = -xTrans / scale;
-            float X = getWidth() / scale;
-
+        if (cropY + Y > mBitmap.getHeight()) {
+            cropY = mBitmap.getHeight() - Y;
             if (DEBUG) {
-                Log.i(TAG, "cropY: " + cropY);
-                Log.i(TAG, "Y: " + Y);
-                Log.i(TAG, "cropX: " + cropX);
-                Log.i(TAG, "X: " + X);
+                Log.i(TAG, "readjust cropY to: " + cropY);
             }
-
-            if (cropY + Y > mBitmap.getHeight()) {
-                cropY = mBitmap.getHeight() - Y;
-                if (DEBUG) {
-                    Log.i(TAG, "readjust cropY to: " + cropY);
-                }
-            }  else if (cropY < 0) {
-                cropY = 0;
-                if (DEBUG) {
-                    Log.i(TAG, "readjust cropY to: " + cropY);
-                }
-            }
-
-            if (cropX + X > mBitmap.getWidth()) {
-                cropX = mBitmap.getWidth() - X;
-                if (DEBUG) {
-                    Log.i(TAG, "readjust cropX to: " + cropX);
-                }
-            } else if (cropX < 0) {
-                cropX = 0;
-                if (DEBUG) {
-                    Log.i(TAG, "readjust cropX to: " + cropX);
-                }
-            }
-
-            if (mBitmap.getHeight() > mBitmap.getWidth()) {
-                // Height is greater than width.
-                if (xTrans >= 0) {
-                    // Image is zoomed. Crop from height
-
-                    Rect src = new Rect(0, (int)cropY, mBitmap.getWidth(), (int)(Y + cropY));
-
-                    if (mAddPaddingToMakeSquare) {
-                        // Crop and add padding to the same bitmap
-                        try {
-                            int size = (int) Y;
-                            bitmap = Bitmap.createBitmap(size, size, mBitmap.getConfig());
-                            Canvas canvas = new Canvas(bitmap);
-                            canvas.drawColor(mPaintColor);
-
-                            // Put cropped bitmap into the canvas
-                            int left = (size - mBitmap.getWidth()) / 2;
-                            Rect dest = new Rect(left, 0, left + mBitmap.getWidth(), size);
-
-                            canvas.drawBitmap(mBitmap, src, dest, null);
-                        } catch (OutOfMemoryError e) {
-                            if (bitmap != null && !bitmap.isRecycled()) {
-                                bitmap.recycle();
-                                bitmap = null;
-                                throw e;
-                            }
-                        }
-
-                    } else {
-
-                        try {
-                            bitmap = Bitmap.createBitmap(mBitmap, 0, (int) cropY, mBitmap.getWidth(), (int) Y,
-                                    null, true);
-                        } catch (OutOfMemoryError e) {
-                            bitmap = null;
-                            throw e;
-                        }
-                    }
-
-                } else {
-                    // Crop from width and height both
-                    try {
-                        bitmap = Bitmap.createBitmap(mBitmap, (int) cropX, (int) cropY, (int) X, (int) Y,
-                                null, true);
-
-                    } catch (OutOfMemoryError e) {
-                        bitmap = null;
-                        throw e;
-                    }
-                }
-            } else {
-                if (yTrans >= 0) {
-                    // Image is zoomed. Crop from width and add padding to make square
-
-                    Rect src = new Rect((int)cropX, 0, (int)(cropX + X), mBitmap.getHeight());
-
-                    if (mAddPaddingToMakeSquare) {
-                        try {
-                            // Crop and add padding to the same bitmap
-                            int size = (int) X;
-                            bitmap = Bitmap.createBitmap(size, size, mBitmap.getConfig());
-                            Canvas canvas = new Canvas(bitmap);
-                            canvas.drawColor(mPaintColor);
-
-                            // Put cropped bitmap into the canvas
-                            int top = (size - mBitmap.getHeight()) / 2;
-                            Rect dest = new Rect(0, top, size, top + mBitmap.getHeight());
-
-                            canvas.drawBitmap(mBitmap, src, dest, null);
-                        } catch (OutOfMemoryError e) {
-                            if (bitmap != null && !bitmap.isRecycled()) {
-                                bitmap.recycle();
-                                bitmap = null;
-                                throw e;
-                            }
-                        }
-                    } else {
-
-                        try {
-                            bitmap = Bitmap.createBitmap(mBitmap, (int) cropX, 0, (int) X, mBitmap.getHeight(),
-                                    null, true);
-                        } catch (OutOfMemoryError e) {
-                            bitmap = null;
-                            throw e;
-                        }
-                    }
-
-                } else {
-                    // Crop from width and height both.
-
-                    try {
-                        bitmap = Bitmap.createBitmap(mBitmap, (int) cropX, (int) cropY, (int) X, (int) Y,
-                                null, true);
-                    } catch (OutOfMemoryError e) {
-                        bitmap = null;
-                        throw e;
-                    }
-
-                }
-
-                if (DEBUG) {
-                    Log.i(TAG, "width should be: " + mBitmap.getWidth());
-                    if (bitmap != null) {
-                        Log.i(TAG, "crop bitmap: " + bitmap.getWidth() + " " + bitmap.getHeight());
-                    }
-                }
+        }  else if (cropY < 0) {
+            cropY = 0;
+            if (DEBUG) {
+                Log.i(TAG, "readjust cropY to: " + cropY);
             }
         }
 
-        return bitmap;
+        if (cropX + X > mBitmap.getWidth()) {
+            cropX = mBitmap.getWidth() - X;
+            if (DEBUG) {
+                Log.i(TAG, "readjust cropX to: " + cropX);
+            }
+        } else if (cropX < 0) {
+            cropX = 0;
+            if (DEBUG) {
+                Log.i(TAG, "readjust cropX to: " + cropX);
+            }
+        }
+
+        Rect rect;
+        int horizontalPadding = 0;
+        int verticalPadding = 0;
+
+        boolean isPaddingRequired = true;
+
+        if (mBitmap.getHeight() > mBitmap.getWidth()) {
+            // Height is greater than width.
+            if (xTrans >= 0) {
+                // Image is zoomed. Crop from height
+                rect = new Rect(0, (int)cropY, mBitmap.getWidth(), (int)(Y + cropY));
+                horizontalPadding = (int) ((Y - mBitmap.getWidth()) / 2);
+            } else {
+                // Crop from width and height both
+                rect = new Rect((int)cropX, (int)cropY, (int)(cropX + X), (int)(cropY + Y));
+                isPaddingRequired = false;
+            }
+        } else {
+            if (yTrans >= 0) {
+                // Image is zoomed. Crop from width and add padding to make square
+                rect = new Rect((int)cropX, 0, (int)(cropX + X), mBitmap.getHeight());
+                verticalPadding = (int) ((X - mBitmap.getHeight()) / 2);
+            } else {
+                // Crop from width and height both.
+                rect = new Rect((int)cropX, (int)cropY, (int)(cropX + X), (int)(cropY + Y));
+                isPaddingRequired = false;
+            }
+        }
+
+        isPaddingRequired = isPaddingRequired && (horizontalPadding != 0 || verticalPadding != 0);
+
+        return CropInfo.cropFromRect(rect, mAddPaddingToMakeSquare && isPaddingRequired, horizontalPadding, verticalPadding, mPaintColor);
+    }
+
+    public Bitmap getCroppedBitmap(CropInfo cropInfo) throws OutOfMemoryError, IllegalArgumentException {
+        if (mBitmap == null) {
+            Log.e(TAG, "original image is not available");
+            return null;
+        }
+
+        return BitmapUtils.getCroppedBitmap(mBitmap, cropInfo);
+    }
+
+    private Bitmap getCroppedBitmap() throws OutOfMemoryError {
+        CropInfo cropInfo = getCropInfo();
+        if (cropInfo != null) {
+            return getCroppedBitmap(cropInfo);
+        }
+
+        return null;
+    }
+
+    public Bitmap getLoadedBitmap() {
+        return mBitmap;
     }
 
     public boolean showAnimation() {
@@ -966,7 +869,6 @@ public class CropperImageView extends ImageView {
     }
 
     private class ScaleListener extends ScaleGestureDetector.SimpleOnScaleGestureListener {
-        protected boolean mScaled = false;
 
         @Override
         public boolean onScale(ScaleGestureDetector detector) {